Add back setup.cfg file
The pyproject metadata stuff depends on
https://peps.python.org/pep-0621/ for which support is only available
in a recent version of setuptools that isn't available in all distributions
just yet. So let's add back a setup.cfg file to allow building on older
distributions.

The tool.setuptools section was also added recently, so we add the
relevant sections for that to setup.cfg as well.